Split disk into mem/ide, added ktest mechanism, initial support for parsing ext2 superblock
This is broken.  Block devices are slightly generalized in block.c now.

ktest is a mechanism to test kernel features that must originate from userland

+++ include/sys/buf.h
@@ -8,12 +8,12 @@
 
 struct buf {
   int flags;
-  uint dev;
-  uint sector;
+  uint8_t dev;
+  uint64_t sector;
   struct buf *prev; // LRU cache list
   struct buf *next;
   struct buf *qnext; // disk queue
-  uchar data[512];
+  uint8_t data[512];
 };
 #define B_BUSY  0x1  // buffer is locked by some process
 #define B_VALID 0x2  // buffer has been read from disk
